+++ command/agent/config.go
@@ -181,7 +181,8 @@ type ClientConfig struct {
 	// Interface to use for network fingerprinting
 	NetworkInterface string `mapstructure:"network_interface"`
 
-	// The network link speed to use if it can not be determined dynamically.
+	// NetworkSpeed is used to override any detected or default network link
+	// speed.
 	NetworkSpeed int `mapstructure:"network_speed"`
 
 	// MaxKillTimeout allows capping the user-specifiable KillTimeout.
@@ -489,7 +490,6 @@ func DefaultConfig() *Config {
 		Vault:          config.DefaultVaultConfig(),
 		Client: &ClientConfig{
 			Enabled:        false,
-			NetworkSpeed:   100,
 			MaxKillTimeout: "30s",
 			ClientMinPort:  14000,
 			ClientMaxPort:  14512,
@@ -533,7 +533,7 @@ func (c *Config) Listener(proto, addr string, port int) (net.Listener, error) {
 			Err: &net.AddrError{Err: "invalid port", Addr: fmt.Sprint(port)},
 		}
 	}
-	return net.Listen(proto, fmt.Sprintf("%s:%d", addr, port))
+	return net.Listen(proto, net.JoinHostPort(addr, strconv.Itoa(port)))
 }
 
 // Merge merges two configurations.
@@ -681,9 +681,9 @@ func (c *Config) normalizeAddrs() error {
 	c.Addresses.RPC = normalizeBind(c.Addresses.RPC, c.BindAddr)
 	c.Addresses.Serf = normalizeBind(c.Addresses.Serf, c.BindAddr)
 	c.normalizedAddrs = &Addresses{
-		HTTP: fmt.Sprintf("%s:%d", c.Addresses.HTTP, c.Ports.HTTP),
-		RPC:  fmt.Sprintf("%s:%d", c.Addresses.RPC, c.Ports.RPC),
-		Serf: fmt.Sprintf("%s:%d", c.Addresses.Serf, c.Ports.Serf),
+		HTTP: net.JoinHostPort(c.Addresses.HTTP, strconv.Itoa(c.Ports.HTTP)),
+		RPC:  net.JoinHostPort(c.Addresses.RPC, strconv.Itoa(c.Ports.RPC)),
+		Serf: net.JoinHostPort(c.Addresses.Serf, strconv.Itoa(c.Ports.Serf)),
 	}
 
 	addr, err := normalizeAdvertise(c.AdvertiseAddrs.HTTP, c.Addresses.HTTP, c.Ports.HTTP, c.DevMode)
@@ -739,7 +739,7 @@ func normalizeAdvertise(addr string, bind string, defport int, dev bool) (string
 			}
 
 			// missing port, append the default
-			return fmt.Sprintf("%s:%d", addr, defport), nil
+			return net.JoinHostPort(addr, strconv.Itoa(defport)), nil
 		}
 		return addr, nil
 	}
@@ -753,11 +753,11 @@ func normalizeAdvertise(addr string, bind string, defport int, dev bool) (string
 	// Return the first unicast address
 	for _, ip := range ips {
 		if ip.IsLinkLocalUnicast() || ip.IsGlobalUnicast() {
-			return fmt.Sprintf("%s:%d", ip, defport), nil
+			return net.JoinHostPort(ip.String(), strconv.Itoa(defport)), nil
 		}
 		if ip.IsLoopback() && dev {
 			// loopback is fine for dev mode
-			return fmt.Sprintf("%s:%d", ip, defport), nil
+			return net.JoinHostPort(ip.String(), strconv.Itoa(defport)), nil
 		}
 	}
 
@@ -776,11 +776,11 @@ func normalizeAdvertise(addr string, bind string, defport int, dev bool) (string
 	// Return the first unicast address
 	for _, ip := range ips {
 		if ip.IsLinkLocalUnicast() || ip.IsGlobalUnicast() {
-			return fmt.Sprintf("%s:%d", ip, defport), nil
+			return net.JoinHostPort(ip.String(), strconv.Itoa(defport)), nil
 		}
 		if ip.IsLoopback() && dev {
 			// loopback is fine for dev mode
-			return fmt.Sprintf("%s:%d", ip, defport), nil
+			return net.JoinHostPort(ip.String(), strconv.Itoa(defport)), nil
 		}
 	}
 	return "", fmt.Errorf("No valid advertise addresses, please set `advertise` manually")